face-recognition/requirements.txt
Flask
face_recognition==1.3.0
numpy==1.21.0
requests==2.25.1
Pillow==8.2.0